+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-17T14:01:25.519Z",
+    "2026-06-18T04:50:35.672Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {